Remove unused imports, rename native_element_info to win32_element_info.
